Move utils to their own module
